@@ -8,10 +8,11 @@
     "@emotion/react": "^11.4.1",
     "@emotion/styled": "^11.3.0",
     "@mui/icons-material": "^5.0.5",
-    "@mui/lab": "^5.0.0-alpha.47",
+    "@mui/lab": "^5.0.0-alpha.84",
     "@mui/material": "^5.0.0",
-    "@mui/styles": "^5.0.0",
+    "@mui/styles": "^5.8.0",
     "@mui/x-data-grid": "^5.2.1",
+    "@mui/x-date-pickers": "^5.0.0-alpha.5",
     "aws-amplify": "^4.3.17",
     "date-fns": "^2.23.0",
     "history": "^5.0.1",
@@ -24,8 +25,8 @@
     "mobx": "^6.2.0",
     "mobx-react-lite": "^3.2.0",
     "notistack": "^2.0.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
     "react-i18next": "^11.13.0",
     "react-router-dom": "^6.0.0",
     "react-scripts": "^5.0.0",
@@ -43,9 +44,11 @@
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
+    "plugins": ["mobx"],
     "extends": [
       "react-app",
-      "react-app/jest"
+      "react-app/jest",
+      "plugin:mobx/recommended"
     ],
     "rules": {
       "@typescript-eslint/indent": [
@@ -79,14 +82,14 @@
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^5.16.3",
-    "@testing-library/react": "^12.1.4",
-    "@testing-library/user-event": "^13.5.0",
-    "@types/jest": "^27.4.1",
+    "@testing-library/react": "^13.3.0",
+    "@testing-library/user-event": "^14.2.0",
+    "@types/jest": "^28.1.1",
     "@types/lodash.clonedeep": "^4.5.6",
     "@types/lodash.merge": "^4.6.6",
     "@types/node": "^17.0.23",
-    "@types/react": "^17.0.43",
-    "@types/react-dom": "^17.0.14",
+    "@types/react": "^18.0.12",
+    "@types/react-dom": "^18.0.5",
     "@types/uuid": "^8.3.4",
     "eslint-config-react-app": "^7.0.0",
     "jest-fetch-mock": "^3.0.3",
