@@ -6,7 +6,7 @@ const mockJudges: Judge[] = [
     district: "Uusimaa",
     email: "tuomari1@sposti.not",
     eventTypes: [],
-    id: 888888,
+    id: 123,
     languages: [],
     location: "Helsinki",
     name: "Tuomari 1",
@@ -17,7 +17,7 @@ const mockJudges: Judge[] = [
     district: "Tampere",
     email: "tuomari2@sposti.not",
     eventTypes: [],
-    id: 999999,
+    id: 223,
     languages: [],
     location: "Tampere",
     name: "Tuomari 2",
@@ -28,7 +28,7 @@ const mockJudges: Judge[] = [
     district: "Turku",
     email: "tuomari3@sposti.not",
     eventTypes: [],
-    id: 777777,
+    id: 333,
     languages: [],
     location: "Turku",
     name: "Tuomari 3",
