@@ -1,5 +1,7 @@
+/* eslint-disable mobx/missing-observer */
 import { CheckBoxOutlineBlank, CheckBox } from "@mui/icons-material";
 import { Autocomplete, AutocompleteProps, Checkbox, TextField } from "@mui/material";
+import { withName } from "./hoc";
 
 type OmitProps = 'disableCloseOnSelect' | 'fullWidth' | 'freeSolo' | 'multiple' | 'renderInput' | 'renderOption' | 'renderTags';
 
@@ -12,9 +14,10 @@ export type AutocompleteMultiProps<T> = Omit<AutocompleteProps<T, true, false, f
 export function AutocompleteMulti<T>(props: AutocompleteMultiProps<T>) {
   const { error, helperText, label, ...acProps } = props;
   const getLabel = props.getOptionLabel || ((o: T) => o);
+  const NamedAutocomplete = withName(Autocomplete<T, true, false, false>, label);
 
   return (
-    <Autocomplete
+    <NamedAutocomplete
       autoHighlight
       data-testid={label}
       {...acProps}
@@ -30,7 +33,7 @@ export function AutocompleteMulti<T>(props: AutocompleteMultiProps<T>) {
             style={{ marginRight: 8 }}
             checked={selected}
           />
-          {getLabel(option)}
+          <>{getLabel(option)}</>
         </li>
       )}
     />
