@@ -1,17 +1,19 @@
 import { TextField } from '@mui/material';
-import { Event } from 'koekalenteri-shared/model';
-import { CollapsibleSection, PartialEvent } from '../..';
+import { AdminEvent } from 'koekalenteri-shared/model';
+import { observer } from 'mobx-react-lite';
+import { CollapsibleSection } from '../..';
+import { CAdminEvent } from '../../../stores/classes';
 
 type AdditionalInfoSectionProps = {
-  event: PartialEvent
-  onChange: (props: Partial<Event>) => void
+  event: CAdminEvent
+  onChange: (props: Partial<AdminEvent>) => void
   onOpenChange?: (value: boolean) => void
   open?: boolean
 }
-export function AdditionalInfoSection({ event, onChange, onOpenChange, open }: AdditionalInfoSectionProps) {
+export const AdditionalInfoSection = observer(function AdditionalInfoSection({ event, onChange, onOpenChange, open }: AdditionalInfoSectionProps) {
   return (
     <CollapsibleSection title="Lisätiedot" open={open} onOpenChange={onOpenChange}>
       <TextField rows={5} fullWidth multiline value={event.description} onChange={(e) => onChange({ description: e.target.value })}></TextField>
     </CollapsibleSection>
   );
-}
+})
