@@ -1,9 +1,11 @@
 import { CheckBox, CheckBoxOutlineBlank } from "@mui/icons-material";
 import { Autocomplete, AutocompleteChangeReason, Avatar, Checkbox, Chip, TextField } from "@mui/material";
-import { isSameDay } from "date-fns";
-import { Event, EventClass, EventState } from "koekalenteri-shared/model";
+import { getDay, isSameDay } from "date-fns";
+import { AdminEvent, EventClass, EventState } from "koekalenteri-shared/model";
+import { runInAction } from "mobx";
+import { observer } from "mobx-react-lite";
 import { useTranslation } from "react-i18next";
-import { PartialEvent } from "../..";
+import { CAdminEvent } from "../../../stores/classes";
 
 /**
  * Callback fired when the value changes.
@@ -21,29 +23,23 @@ type EventClassesOnChange = (
 
 type EventClassesProps = {
   id: string
-  event: PartialEvent
+  event: CAdminEvent
   value: EventClass[] | undefined
   classes: EventClass[]
   label: string
   required?: boolean
   requiredState?: EventState
-  errorStates?: { [Property in keyof Event]?: boolean }
-  helperTexts?: { [Property in keyof Event]?: string }
+  errorStates?: { [Property in keyof AdminEvent]?: boolean }
+  helperTexts?: { [Property in keyof AdminEvent]?: string }
   onChange: EventClassesOnChange
 }
 
-export const compareEventClass = (a: EventClass, b: EventClass) =>
-  isSameDay(a.date || new Date(), b.date || new Date())
-    ? a.class.localeCompare(b.class)
-    : (a.date?.valueOf() || 0) - (b.date?.valueOf() || 0);
+export const isSameEventClass = (a: EventClass, b: EventClass) =>
+  isSameDay(a.date || new Date(), b.date || new Date()) && a.class === b.class;
 
-export function EventClasses(props: EventClassesProps) {
-  if (props.value) {
-    props.value.sort(compareEventClass);
-  }
-
+export const EventClasses = observer(function EventClasses(props: EventClassesProps) {
   const { t } = useTranslation();
-  const { classes, label, event, required, requiredState, errorStates, helperTexts, ...rest } = props;
+  const { classes, label, event, required, requiredState, errorStates, helperTexts, onChange, ...rest } = props;
   const error = errorStates?.classes;
   const helperText = helperTexts?.classes || '';
 
@@ -58,7 +54,12 @@ export function EventClasses(props: EventClassesProps) {
       groupBy={c => t('weekday', { date: c.date })}
       options={classes}
       getOptionLabel={c => c.class}
-      isOptionEqualToValue={(o, v) => compareEventClass(o, v) === 0}
+      isOptionEqualToValue={(o, v) => isSameEventClass(o, v)}
+      onChange={(e, v, r) => {
+        if (onChange) {
+          runInAction(() => onChange(e, v, r))
+        }
+      }}
       renderOption={(optionProps, option, { selected }) => (
         <li {...optionProps}>
           <Checkbox
@@ -74,23 +75,21 @@ export function EventClasses(props: EventClassesProps) {
       renderTags={(tagValue, getTagProps) => tagValue.map((option, index) => (
         <Chip
           {...getTagProps({ index })}
-          avatar={
-            <Avatar
-              sx={{
-                fontWeight: 'bold',
-                bgcolor: isSameDay(option.date || event.startDate, event.startDate) ? 'secondary.light' : 'secondary.dark'
-              }}
-            >
-              {t('weekday', { date: option.date })}
-            </Avatar>
-          }
+          avatar={<Avatar>{t('weekday', { date: option.date })}</Avatar>}
           label={option.class}
           onDelete={undefined}
           size="small"
-          sx={{bgcolor: option.judge?.id ? 'background.ok' : 'transparent'}}
+          sx={{
+            bgcolor: option.judge?.id ? 'background.ok' : 'transparent',
+            '> .MuiChip-avatar': {
+              color: 'white',
+              fontWeight: 'bold',
+              bgcolor: 'background.weekdays.'+getDay(option.date || event.startDate)
+            }
+          }}
           variant={option.judge ? "filled" : "outlined"}
         />
       ))}
     />
   );
-}
+})
