@@ -1,29 +1,20 @@
 import { Checkbox, FormControlLabel, FormGroup, Grid, Switch, TextField } from '@mui/material';
-import { Registration, RegistrationPerson } from 'koekalenteri-shared/model';
+import { action } from 'mobx';
+import { observer } from 'mobx-react-lite';
 import { useTranslation } from 'react-i18next';
-import { CollapsibleSection, emptyPerson } from '../..';
-import { useStores } from '../../../stores';
+import { CollapsibleSection } from '../..';
+import { CRegistration } from '../../../stores/classes/CRegistration';
 
 type OwnerInfoProps = {
-  reg: Partial<Registration>
+  reg: CRegistration
   error?: boolean
   helperText?: string
-  onChange: (props: Partial<Registration>) => void
   onOpenChange?: (value: boolean) => void
   open?: boolean
 };
 
-export function OwnerInfo({reg, error, helperText, onChange, onOpenChange, open}: OwnerInfoProps) {
+export const OwnerInfo = observer(function OwnerInfo({reg, error, helperText, onOpenChange, open}: OwnerInfoProps) {
   const { t } = useTranslation();
-  const { rootStore } = useStores();
-
-  const handleChange = (props: Partial<RegistrationPerson>) => {
-    const owner = { ...emptyPerson, ...reg.owner, ...props };
-    if (reg.dog?.regNo) {
-      rootStore.dogStore.save({ dog: { ...reg.dog }, owner });
-    }
-    onChange({ owner });
-  }
 
   return (
     <CollapsibleSection title={t('registration.owner')} error={error} helperText={helperText} open={open} onOpenChange={onOpenChange}>
@@ -32,13 +23,13 @@ export function OwnerInfo({reg, error, helperText, onChange, onOpenChange, open}
           <Grid item sx={{ width: 300 }}>
             <TextField
               InputProps={{ autoComplete: 'name' }}
-              error={!reg.owner?.name}
+              error={!reg.owner.name}
               fullWidth
               id="owner_name"
               label={t('registration.contact.name')}
               name="name"
-              onChange={e => handleChange({ name: e.target.value || '' })}
-              value={reg.owner?.name || ''}
+              onChange={action(e => { reg.owner = { ...reg.owner, name: e.target.value } })}
+              value={reg.owner.name || ''}
             />
           </Grid>
           <Grid item sx={{ width: 300 }}>
@@ -49,8 +40,8 @@ export function OwnerInfo({reg, error, helperText, onChange, onOpenChange, open}
               id="owner_city"
               label={t('registration.contact.city')}
               name="city"
-              onChange={e => handleChange({ location: e.target.value || '' })}
-              value={reg.owner?.location || ''}
+              onChange={action(e => { reg.owner = { ...reg.owner, location: e.target.value } })}
+              value={reg.owner.location || ''}
             />
           </Grid>
         </Grid>
@@ -63,8 +54,8 @@ export function OwnerInfo({reg, error, helperText, onChange, onOpenChange, open}
               id="owner_email"
               label={t('registration.contact.email')}
               name="email"
-              onChange={e => handleChange({ email: e.target.value || '' })}
-              value={reg.owner?.email || ''}
+              onChange={action(e => { reg.owner = { ...reg.owner, email: e.target.value } })}
+              value={reg.owner.email || ''}
             />
           </Grid>
           <Grid item sx={{ width: 300 }}>
@@ -75,8 +66,8 @@ export function OwnerInfo({reg, error, helperText, onChange, onOpenChange, open}
               id="owner_phone"
               label={t('registration.contact.phone')}
               name="phone"
-              onChange={e => handleChange({ phone: e.target.value || '' })}
-              value={reg.owner?.phone || ''}
+              onChange={action(e => { reg.owner = { ...reg.owner, phone: e.target.value } })}
+              value={reg.owner.phone || ''}
             />
           </Grid>
         </Grid>
@@ -84,22 +75,19 @@ export function OwnerInfo({reg, error, helperText, onChange, onOpenChange, open}
       <FormGroup>
         <FormControlLabel control={
           <Checkbox
-            checked={reg.owner?.membership}
-            onChange={e => handleChange({ membership: e.target.checked })}
+            checked={!!reg.owner?.membership}
+            onChange={e => { reg.owner = { ...reg.owner, membership: e.target.checked } }}
           />} label={t('registration.ownerIsMember')}
         />
       </FormGroup>
       <FormGroup>
         <FormControlLabel control={
           <Switch
-            checked={reg.ownerHandles}
-            onChange={e => onChange({
-              ownerHandles: e.target.checked,
-              handler: e.target.checked ? { ...emptyPerson, ...reg.owner } : { ...emptyPerson }
-            })}
+            checked={!!reg.ownerHandles}
+            onChange={action(e => { reg.ownerHandles = e.target.checked })}
           />
         } label={t('registration.ownerHandles')} />
       </FormGroup>
     </CollapsibleSection>
   );
-}
+})
