@@ -1,29 +1,20 @@
 import { Checkbox, FormControlLabel, Grid, TextField } from '@mui/material';
-import { Registration, RegistrationPerson } from 'koekalenteri-shared/model';
+import { action } from 'mobx';
+import { observer } from 'mobx-react-lite';
 import { useTranslation } from 'react-i18next';
-import { CollapsibleSection, emptyPerson } from '../..';
-import { useStores } from '../../../stores';
+import { CollapsibleSection } from '../..';
+import { CRegistration } from '../../../stores/classes/CRegistration';
 
 type HandlerInfoProps = {
-  reg: Partial<Registration>
+  reg: CRegistration
   error?: boolean
   helperText?: string
-  onChange: (props: Partial<Registration>) => void
   onOpenChange?: (value: boolean) => void
   open?: boolean
 };
 
-export function HandlerInfo({ reg, error, helperText, onChange, onOpenChange, open }: HandlerInfoProps) {
+export const HandlerInfo = observer(function HandlerInfo({ reg, error, helperText, onOpenChange, open }: HandlerInfoProps) {
   const { t } = useTranslation();
-  const { rootStore } = useStores();
-
-  const handleChange = (props: Partial<RegistrationPerson>) => {
-    const handler = { ...emptyPerson, ...reg.handler, ...props };
-    if (reg.dog?.regNo) {
-      rootStore.dogStore.save({ dog: { ...reg.dog }, handler });
-    }
-    onChange({ handler });
-  }
 
   return (
     <CollapsibleSection title={t('registration.handler')} error={error} helperText={helperText} open={open} onOpenChange={onOpenChange}>
@@ -37,8 +28,8 @@ export function HandlerInfo({ reg, error, helperText, onChange, onOpenChange, op
               id="handler_name"
               label={t('registration.contact.name')}
               name="name"
-              onChange={e => handleChange({ name: e.target.value || '' })}
-              value={reg.handler?.name || ''}
+              onChange={action(e => { reg.handler = { ...reg.handler, name: e.target.value } })}
+              value={reg.handler.name || ''}
             />
           </Grid>
           <Grid item sx={{ width: 300 }}>
@@ -48,7 +39,7 @@ export function HandlerInfo({ reg, error, helperText, onChange, onOpenChange, op
               fullWidth
               id="handler_city"
               name="city"
-              onChange={e => handleChange({ location: e.target.value || '' })}
+              onChange={action(e => { reg.handler = { ...reg.handler, location: e.target.value } })}
               label={t('registration.contact.city')}
               value={reg.handler?.location || ''}
             />
@@ -63,7 +54,7 @@ export function HandlerInfo({ reg, error, helperText, onChange, onOpenChange, op
               id="handler_email"
               label={t('registration.contact.email')}
               name="email"
-              onChange={e => handleChange({ email: e.target.value || '' })}
+              onChange={action(e => { reg.handler = { ...reg.handler, email: e.target.value } })}
               value={reg.handler?.email || ''}
             />
           </Grid>
@@ -75,7 +66,7 @@ export function HandlerInfo({ reg, error, helperText, onChange, onOpenChange, op
               id="handler_phone"
               label={t('registration.contact.phone')}
               name="phone"
-              onChange={e => handleChange({ phone: e.target.value || '' })}
+              onChange={action(e => { reg.handler = { ...reg.handler, phone: e.target.value } })}
               value={reg.handler?.phone || ''}
             />
           </Grid>
@@ -84,12 +75,12 @@ export function HandlerInfo({ reg, error, helperText, onChange, onOpenChange, op
       <FormControlLabel
         control={
           <Checkbox
-            checked={reg.handler?.membership}
-            onChange={e => handleChange({ membership: e.target.checked })}
+            checked={!!reg.handler?.membership}
+            onChange={action(e => { reg.handler = { ...reg.handler, membership: e.target.checked } })}
           />
         }
         label={t('registration.handlerIsMember')}
       />
     </CollapsibleSection>
   );
-}
+})
