@@ -1,61 +1,66 @@
 import { Box, CircularProgress, Grid, Paper, Toolbar, Typography } from '@mui/material';
-import type { ConfirmedEventEx, Registration } from 'koekalenteri-shared/model';
+import { autorun, runInAction } from 'mobx';
 import { observer } from 'mobx-react-lite';
 import { useSnackbar } from 'notistack';
-import { useEffect, useState } from 'react';
+import { useEffect } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useNavigate, useParams } from 'react-router-dom';
-import { putRegistration } from '../api/event';
 import { CollapsibleSection, CostInfo, LinkButton, RegistrationForm } from '../components';
+import { getRegistrationDates } from '../components/forms/RegistrationForm/1.Entry';
 import { Header } from '../layout';
 import { useSessionStarted, useStores } from '../stores';
+import { CEvent } from '../stores/classes';
+import { CRegistration } from '../stores/classes/CRegistration';
 import { entryDateColor } from '../utils';
 
 export const EventRegistrationPage = observer(function EventRegistrationPage() {
   const params = useParams();
-  const { publicStore } = useStores();
-  const [event, setEvent] = useState<ConfirmedEventEx>();
+  const { rootStore } = useStores();
   const [sessionStarted] = useSessionStarted();
   const { t } = useTranslation();
 
   useEffect(() => {
+    if (!rootStore.loaded) {
+      rootStore.load();
+    }
+  }, []);
+
+  useEffect(() => autorun(() => {
     const abort = new AbortController();
     async function get(eventType: string, id: string) {
-      const result = await publicStore.get(eventType, id, abort.signal) as ConfirmedEventEx;
-      setEvent(result);
+      rootStore.eventStore.selectedEvent = await rootStore.eventStore.get(eventType, id, abort.signal);
     }
-    if (params.eventType && params.id) {
+    if (params.eventType && params.id && rootStore.eventStore.selectedEvent?.id !== params.id) {
       get(params.eventType, params.id);
     }
     return () => abort.abort();
-  }, [params, publicStore]);
+  // eslint-disable-next-line react-hooks/exhaustive-deps
+  }), [params]);
 
   return (
     <>
-      <Header title={t('entryTitle', { context: event?.eventType === 'other' ? '' : 'test' })} />
+      <Header title={t('entryTitle', { context: rootStore.eventStore.selectedEvent?.eventType === 'other' ? '' : 'test' })} />
       <Box sx={{ p: 1, overflow: 'hidden', height: '100%', display: 'flex', flexDirection: 'column', alignItems: 'flex-start' }}>
         <Toolbar variant="dense" />{/* To allocate the space for fixed header */}
         <LinkButton sx={{mb: 1}} to="/" text={sessionStarted ? t('goBack') : t('goHome')} />
-        {event ? <EventComponent event={event} classDate={params.date} className={params.class} /> : <CircularProgress />}
+        {rootStore.eventStore.selectedEvent ? <EventComponent event={rootStore.eventStore.selectedEvent} classDate={params.date} className={params.class} /> : <CircularProgress />}
       </Box>
     </>
   )
 })
 
-function EventComponent({ event, classDate = '', className = '' }: { event: ConfirmedEventEx, classDate?: string, className?: string }) {
+const EventComponent = observer(function EventComponent({ event, classDate = '', className = '' }: { event: CEvent, classDate?: string, className?: string }) {
   const { enqueueSnackbar } = useSnackbar();
-  const { publicStore } = useStores();
+  const { rootStore } = useStores();
   const { t } = useTranslation();
   const navigate = useNavigate();
-  const onSave = async (registration: Registration) => {
+  const onSave = async (registration: CRegistration) => {
     try {
-      const saved = await putRegistration(registration);
-      console.log(saved);
-      publicStore.load(); // TODO: Use MobX properly
+      const result = await rootStore.registrationStore.save(registration);
       navigate('/', { replace: true });
-      const emails = [saved.handler.email];
-      if (saved.owner.email !== saved.handler.email) {
-        emails.push(saved.owner.email);
+      const emails = [result.handler.email];
+      if (result.owner.email !== result.handler.email) {
+        emails.push(result.owner.email);
       }
       enqueueSnackbar(t('registration.saved', { count: emails.length, to: emails.join("\n") }), { variant: 'success', style: { whiteSpace: 'pre-line' } });
       return true;
@@ -68,10 +73,18 @@ function EventComponent({ event, classDate = '', className = '' }: { event: Conf
     navigate('/');
     return true;
   }
+  useEffect(() => runInAction(() => {
+    rootStore.registrationStore.registration.eventId = event.id;
+    rootStore.registrationStore.registration.eventType = event.eventType;
+    if (className) {
+      rootStore.registrationStore.registration.class = className;
+    }
+    rootStore.registrationStore.registration.dates = getRegistrationDates(event, classDate, className);
+  }));
   return (
     <>
       <Paper sx={{ p: 1, mb: 1, width: '100%' }} elevation={2}>
-        <Typography variant="h5" sx={{fontWeight: 'bold', mb: 1}}>
+        <Typography variant="h5" sx={{ fontWeight: 'bold', mb: 1 }}>
           {event.eventType} {t('daterange', { start: event.startDate, end: event.endDate })} {event.location + (event.name ? ` (${event.name})` : '')}
         </Typography>
         <CollapsibleSection title="Kokeen tiedot">
@@ -80,14 +93,14 @@ function EventComponent({ event, classDate = '', className = '' }: { event: Conf
           >
             <Grid item container sm={12} md columnSpacing={1}>
               <Grid item xs={4}>{t('entryTime')}:</Grid>
-              <Grid item xs={8} sx={{ color: entryDateColor(event) }}>
+              <Grid item xs={8} sx={{ color: entryDateColor(event.isEntryOpen, event.isEntryClosing) }}>
                 <b>{t('daterange', { start: event.entryStartDate, end: event.entryEndDate })}</b>&nbsp;
                 {event.isEntryOpen ? t('distanceLeft', { date: event.entryEndDate }) : ''}
               </Grid>
               <Grid item xs={4}>{t('event.organizer')}:</Grid>
               <Grid item xs={8}>{event.organizer?.name}</Grid>
               <Grid item xs={4}>{t('event.judges')}:</Grid>
-              <Grid item xs={8}>{event.judges[0]}</Grid>
+              <Grid item xs={8}>{event.judges.length > 0 && event.judges[0]?.name}</Grid>
               <Grid item xs={4}>{t('event.official')}:</Grid>
               <Grid item xs={8}>{event.official?.name || ''}</Grid>
             </Grid>
@@ -100,7 +113,14 @@ function EventComponent({ event, classDate = '', className = '' }: { event: Conf
           </Grid>
         </CollapsibleSection>
       </Paper>
-      <RegistrationForm event={event} className={className} classDate={classDate} onSave={onSave} onCancel={onCancel}/>
+      <RegistrationForm
+        event={event}
+        registration={rootStore.registrationStore.registration}
+        className={className}
+        classDate={classDate}
+        onSave={onSave}
+        onCancel={onCancel}
+      />
     </>
   );
-}
+})
