@@ -1,8 +1,9 @@
+/* eslint-disable mobx/missing-observer */
 import { Authenticator, useAuthenticator } from '@aws-amplify/ui-react';
 import { useSnackbar } from 'notistack';
 import { useEffect } from 'react';
 import { useLocation, useNavigate } from 'react-router-dom';
-import { useSessionBoolean, useSessionStorage } from '../stores';
+import { useSessionBoolean } from '../stores';
 
 export function LoginPage() {
   const { user, route } = useAuthenticator(context => [context.user, context.route]);
@@ -11,13 +12,6 @@ export function LoginPage() {
   const location = useLocation();
   const navigate = useNavigate();
   const state: Record<string, any> = location.state as any || {};
-  const [from, setFrom] = useSessionStorage('auth-from', state.from?.pathname || '/');
-
-  useEffect(() => {
-    if (state && state.from && state.from.pathname !== from) {
-      setFrom(state.from.pathname);
-    }
-  });
 
   useEffect(() => {
     if (route === 'authenticated') {
@@ -25,7 +19,7 @@ export function LoginPage() {
         enqueueSnackbar(`Tervetuloa, ${user.attributes?.name || user.attributes?.email}!`, { variant: 'info' });
         setGreeted(true);
       } else {
-        navigate(from || '/', { replace: true });
+        navigate(state?.from?.pathname || '/', { replace: true });
       }
     }
   });
