@@ -1,27 +1,26 @@
 import { Box } from '@mui/material';
-import { toJS } from 'mobx';
 import { observer } from 'mobx-react-lite';
 import { useEffect, useState } from 'react';
 import { useSearchParams } from 'react-router-dom';
-import { deserializeFilter, EventFilter, serializeFilter } from '../components';
-import { Banner, EventContainer, Header } from '../layout';
+import { deserializeFilter, EventFilter, EventTable, serializeFilter } from '../components';
+import { Banner, Header } from '../layout';
 import { useSessionStarted, useStores } from '../stores';
-import { FilterProps } from '../stores/PublicStore';
+import { FilterProps } from '../stores/EventStore';
 
 export const SearchPage = observer(function SearchPage() {
-  const { rootStore, publicStore } = useStores();
+  const { rootStore } = useStores();
   const [sessionStarted, setSessionStarted] = useSessionStarted();
   const [searchParams, setSearchParams] = useSearchParams();
   const [filter, setFilter] = useState(deserializeFilter(searchParams));
 
-  const organizers = toJS(rootStore.organizerStore.organizers);
-  const judges = toJS(rootStore.judgeStore.activeJudges);
+  const organizers = rootStore.organizerStore.organizers.map(o => o.toJSON());
+  const judges = rootStore.judgeStore.activeJudges.map(j => j.toJSON());
   const eventTypes = rootStore.eventTypeStore.activeEventTypes.map(et => et.eventType);
 
   const handleChange = (filter: FilterProps) => {
     setFilter(filter);
     setSearchParams(serializeFilter(filter));
-    publicStore.setFilter(filter);
+    rootStore.eventStore.setFilter(filter);
   };
 
   useEffect(() => {
@@ -31,16 +30,13 @@ export const SearchPage = observer(function SearchPage() {
     if (!rootStore.loaded) {
       rootStore.load();
     }
-    if (!publicStore.loaded) {
-      publicStore.load();
-    }
   });
 
   useEffect(() => {
     const value = deserializeFilter(searchParams);
     setFilter(value);
-    publicStore.setFilter(value);
-  }, [searchParams, publicStore]);
+    rootStore.eventStore.setFilter(value);
+  }, [searchParams, rootStore.eventStore]);
 
   return (
     <>
@@ -48,7 +44,7 @@ export const SearchPage = observer(function SearchPage() {
       <Banner />
       <Box>
         <EventFilter organizers={organizers} judges={judges} filter={filter} eventTypes={eventTypes} onChange={handleChange} />
-        <EventContainer />
+        <EventTable events={rootStore.eventStore.filteredEvents} loading={rootStore.loading} />
       </Box>
     </>
   )
