@@ -1,15 +1,15 @@
 import { useAuthenticator } from '@aws-amplify/ui-react';
-import { Box, Toolbar } from '@mui/material';
-import { autorun } from 'mobx';
+import { Box, CircularProgress, Toolbar } from '@mui/material';
+import { observer } from 'mobx-react-lite';
 import { ReactNode, useEffect, useState } from 'react';
 import { Navigate, useLocation } from 'react-router-dom';
 import { Header, SideMenu } from '../../layout';
 import { useSessionStarted, useStores } from '../../stores';
 
-export function AuthPage({ children, title }: { children: ReactNode, title?: string }) {
+export const AuthPage = observer(function AuthPage({ children, title }: { children: ReactNode, title?: string }) {
   const location = useLocation();
   const { route } = useAuthenticator(context => [context.route]);
-  const { rootStore, publicStore, privateStore } = useStores();
+  const { rootStore } = useStores();
   const [menuOpen, setMenuOpen] = useState(false);
   const [sessionStarted, setSessionStarted] = useSessionStarted();
 
@@ -17,37 +17,21 @@ export function AuthPage({ children, title }: { children: ReactNode, title?: str
     if (!sessionStarted) {
       setSessionStarted(new Date().toISOString());
     }
-  });
-  useEffect(() => {
     if (!rootStore.loaded) {
       rootStore.load();
     }
-    autorun(() => {
-      if (!publicStore.loaded) {
-        publicStore.load();
-      }
-      if (!privateStore.loaded) {
-        privateStore.load();
-      }
-    })
   });
-
-
-  const toggleMenu = () => {
-    console.log('toggleMenu', menuOpen);
-    setMenuOpen(!menuOpen);
-  }
 
   return (route !== 'authenticated' ? <Navigate to="/login" state={{ from: location }} replace /> :
     <>
-      <Header title={title} toggleMenu={toggleMenu} />
+      <Header title={title} toggleMenu={() => setMenuOpen(!menuOpen)} />
       <Box sx={{ display: 'flex', height: '100%' }}>
         <SideMenu open={menuOpen} onClose={() => setMenuOpen(false)} />
         <Box sx={{ p: 1, display: 'flex', flexDirection: 'column', flexGrow: 1, overflow: 'auto' }}>
           <Toolbar variant="dense" />
-          {children}
+          {rootStore.loading ? <CircularProgress /> : children}
         </Box>
       </Box>
     </>
   );
-};
+})
