@@ -1,9 +1,7 @@
+import { configure } from "mobx";
 import { createContext, useContext, useMemo } from "react";
 import { Setter, useLocalStorage, useSessionStorage } from './browser';
-import { PublicStore } from "./PublicStore";
-import { PrivateStore } from "./PrivateStore";
 import { RootStore } from "./RootStore";
-import { configure } from "mobx";
 
 // Make sure TS is configured properly for MobX
 if (!new class { x: any }().hasOwnProperty('x')) throw new Error('Transpiler is not configured correctly');
@@ -12,15 +10,14 @@ if (!new class { x: any }().hasOwnProperty('x')) throw new Error('Transpiler is 
 const notInTest = process.env.NODE_ENV !== 'test';
 configure({
   enforceActions: "always",
-  computedRequiresReaction: notInTest,
+  computedRequiresReaction: false, //notInTest,
   reactionRequiresObservable: notInTest,
-  observableRequiresReaction: notInTest
+  observableRequiresReaction: false,
+  safeDescriptors: true
 });
 
 const rootStoreContext = createContext({
   rootStore: new RootStore(),
-  publicStore: new PublicStore(),
-  privateStore: new PrivateStore(),
 });
 
 export const useStores = () => useContext(rootStoreContext);
@@ -36,3 +33,4 @@ export const useSessionBoolean = (key: string, defaultValue: boolean): [boolean,
 }
 
 export { useLocalStorage, useSessionStorage };
+
