@@ -1,10 +1,13 @@
-import type { EventEx } from 'koekalenteri-shared/model';
+import { parseISO, startOfDay } from "date-fns";
 
-export function entryDateColor(event: EventEx) {
-  if (!event.isEntryOpen) {
+export function entryDateColor(isEntryOpen: boolean, isEntryClosing: boolean, extended?: boolean) {
+  if (!isEntryOpen) {
     return 'text.primary';
   }
-  return event.isEntryClosing ? 'warning.main' : 'success.main';
+  if (extended) {
+    return 'error.main';
+  }
+  return isEntryClosing ? 'warning.main' : 'success.main';
 }
 
 export function unique(arr: string[]): string[] {
@@ -13,4 +16,12 @@ export function unique(arr: string[]): string[] {
 
 export function uniqueDate(arr: Date[]): Date[] {
   return arr.filter((c, i, a) => a.findIndex(f => f.valueOf() === c.valueOf()) === i);
+}
+
+export function toDate(value: string | undefined): Date | undefined {
+  return value ? parseISO(value) : undefined;
+}
+
+export function toDateOrNow(value: string | undefined): Date {
+  return startOfDay(value ? parseISO(value) : new Date());
 }
